Renamed common_widgets folder to just widgets
