+++ packages/uni_app/lib/view/about/about.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:uni/view/about/widgets/terms_and_conditions.dart';
-import 'package:uni/view/common_widgets/pages_layouts/secondary/secondary.dart';
+import 'package:uni/view/widgets/pages_layouts/secondary/secondary.dart';
 
 class AboutPageView extends StatefulWidget {
   const AboutPageView({super.key});
