+++ packages/uni_app/lib/view/academic_path/widgets/no_courses_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:uni/generated/l10n.dart';
-import 'package:uni/view/common_widgets/expanded_image_label.dart';
+import 'package:uni/view/widgets/expanded_image_label.dart';
 
 class NoCoursesWidget extends StatelessWidget {
   const NoCoursesWidget({super.key});
