+++ packages/uni_app/lib/view/course_unit_info/widgets/course_unit_info_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:uni/view/common_widgets/generic_expansion_card.dart';
+import 'package:uni/view/widgets/generic_expansion_card.dart';
 
 class CourseUnitInfoCard extends GenericExpansionCard {
   const CourseUnitInfoCard(this.sectionTitle, this.content, {super.key})
