+++ packages/uni_app/lib/view/course_unit_info/widgets/course_unit_sheet.dart
@@ -10,8 +10,8 @@ import 'package:uni/model/entities/course_units/sheet.dart';
 import 'package:uni/model/entities/exam.dart';
 import 'package:uni/model/providers/startup/profile_provider.dart';
 import 'package:uni/model/providers/startup/session_provider.dart';
-import 'package:uni/view/common_widgets/generic_animated_expandable.dart';
-import 'package:uni/view/common_widgets/generic_expandable.dart';
+import 'package:uni/view/widgets/generic_animated_expandable.dart';
+import 'package:uni/view/widgets/generic_expandable.dart';
 import 'package:uni/view/course_unit_info/widgets/modal_professor_info.dart';
 import 'package:uni_ui/cards/book_card.dart';
 import 'package:uni_ui/cards/exam_card.dart';
