+++ packages/uni_app/lib/view/faculty/faculty.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:uni/generated/l10n.dart';
 import 'package:uni/model/providers/lazy/library_occupation_provider.dart';
 import 'package:uni/utils/navigation_items.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/general.dart';
+import 'package:uni/view/widgets/pages_layouts/general/general.dart';
 import 'package:uni/view/faculty/widgets/service_cards.dart';
 import 'package:uni/view/home/widgets/calendar/calendar_home_card.dart';
 import 'package:uni/view/home/widgets/library/library_home_card.dart';
