+++ packages/uni_app/lib/view/home/widgets/exams/exam_home_card.dart
@@ -7,7 +7,7 @@ import 'package:uni/model/providers/lazy/exam_provider.dart';
 import 'package:uni/utils/date_time_formatter.dart';
 import 'package:uni/utils/string_formatter.dart';
 import 'package:uni/view/academic_path/academic_path.dart';
-import 'package:uni/view/common_widgets/icon_label.dart';
+import 'package:uni/view/widgets/icon_label.dart';
 import 'package:uni/view/home/widgets/generic_home_card.dart';
 import 'package:uni/view/home/widgets/schedule/timeline_shimmer.dart';
 import 'package:uni/view/lazy_consumer.dart';
