+++ packages/uni_app/lib/view/profile/widgets/profile_overview.dart
@@ -3,7 +3,7 @@ import 'package:provider/provider.dart';
 import 'package:uni/model/entities/profile.dart';
 import 'package:uni/model/providers/startup/profile_provider.dart';
 import 'package:uni/model/providers/startup/session_provider.dart';
-import 'package:uni/view/common_widgets/profile_image.dart';
+import 'package:uni/view/widgets/profile_image.dart';
 
 class ProfileOverview extends StatelessWidget {
   const ProfileOverview({
