+++ packages/uni_app/lib/view/widgets/pages_layouts/general/widgets/bottom_navigation_bar.dart
