+++ packages/uni_app/lib/view/widgets/pages_layouts/secondary/secondary.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/general.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/widgets/bottom_navigation_bar.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/widgets/refresh_state.dart';
-import 'package:uni/view/common_widgets/pages_layouts/general/widgets/top_navigation_bar.dart';
+import 'package:uni/view/widgets/pages_layouts/general/general.dart';
+import 'package:uni/view/widgets/pages_layouts/general/widgets/bottom_navigation_bar.dart';
+import 'package:uni/view/widgets/pages_layouts/general/widgets/refresh_state.dart';
+import 'package:uni/view/widgets/pages_layouts/general/widgets/top_navigation_bar.dart';
 
 /// Page with a back button on top
 abstract class SecondaryPageViewState<T extends StatefulWidget>
